+++ src/controls/text.example.ts
@@ -3,7 +3,7 @@ import WUPTextControl from "./text";
 
 // it's required to extend default validationRules
 declare global {
-  namespace WUPText {
+  namespace WUP.Text {
     interface ValidationMap {
       isNumber: boolean;
     }