+++ src/popup/popupListen.ts
@@ -2,7 +2,7 @@ import focusFirst from "../helpers/focusFirst";
 import onEvent, { onEventType } from "../helpers/onEvent";
 import onFocusGot from "../helpers/onFocusGot";
 import onFocusLost from "../helpers/onFocusLost";
-import { WUPPopup } from "./popupElement.types";
+import { HideCases, ShowCases } from "./popupElement.types";
 
 /** Listen for target according to showCase. If target removed call stopListen + don't forget to remove popup yourself;
  * @tutorial Troubleshooting/rules:
@@ -11,7 +11,7 @@ import { WUPPopup } from "./popupElement.types";
 export default function popupListen(
   options: {
     target: HTMLElement;
-    showCase?: WUPPopup.ShowCases;
+    showCase?: ShowCases;
     hoverShowTimeout?: number;
     hoverHideTimeout?: number;
     focusDebounceMs?: number;
@@ -19,18 +19,15 @@ export default function popupListen(
     skipAlreadyFocused?: boolean;
   },
   /** If succesfull callback must return HTMLElement */
-  onShow: (
-    showCase: WUPPopup.ShowCases,
-    ev: MouseEvent | FocusEvent | null
-  ) => HTMLElement | null | Promise<HTMLElement | null>,
-  onHide: (hideCase: WUPPopup.HideCases, ev: MouseEvent | FocusEvent | null) => boolean | Promise<boolean>
+  onShow: (showCase: ShowCases, ev: MouseEvent | FocusEvent | null) => HTMLElement | null | Promise<HTMLElement | null>,
+  onHide: (hideCase: HideCases, ev: MouseEvent | FocusEvent | null) => boolean | Promise<boolean>
 ): {
   /** Fire it when target is removed manually (to remove all added related eventListeners) */
   stopListen: () => void;
   /** Fire it before hiding (to ability to show in the feature by listened events) when you need to hide outside listener (manually) */
-  hide: (hideCase: WUPPopup.HideCases) => Promise<void>;
+  hide: (hideCase: HideCases) => Promise<void>;
   /** Fire it before showing (to ability to hide in the feature by listened events) when you need to show outside listener (manually) */
-  show: (showCase: WUPPopup.ShowCases) => Promise<void>;
+  show: (showCase: ShowCases) => Promise<void>;
 } {
   const opts = { ...popupListen.$defaults, ...options };
   const t = opts.target;
@@ -59,7 +56,7 @@ export default function popupListen(
     onShowCallbacks.length = 0;
   }
 
-  async function show(showCase: WUPPopup.ShowCases, e?: MouseEvent | FocusEvent): Promise<void> {
+  async function show(showCase: ShowCases, e?: MouseEvent | FocusEvent): Promise<void> {
     if (openedEl || show._isDoing) {
       return;
     }
@@ -79,7 +76,7 @@ export default function popupListen(
   }
   show._isDoing = false as Promise<void> | boolean; // required to prevent infinite-calling from parent
 
-  async function hide(hideCase: WUPPopup.HideCases, e?: MouseEvent | FocusEvent | null): Promise<void> {
+  async function hide(hideCase: HideCases, e?: MouseEvent | FocusEvent | null): Promise<void> {
     show._isDoing && (await show._isDoing);
     if (!openedEl || hide._isDoing) {
       return;
@@ -125,7 +122,7 @@ export default function popupListen(
   let openedByHover = false;
   let debounceTimeout: ReturnType<typeof setTimeout> | undefined;
   // onClick
-  if (opts.showCase & WUPPopup.ShowCases.onClick) {
+  if (opts.showCase & ShowCases.onClick) {
     let wasOutsideClick = false; // fix when labelOnClick > inputOnClick
     let lastActive: HTMLElement | null = null;
     onShowEvent(document, "focusin", ({ target }) => {
@@ -164,9 +161,9 @@ export default function popupListen(
         const isMeClick = openedEl === e.target || includes(e.target);
         if (isMeClick) {
           focusFirst(lastActive || t);
-          hide(WUPPopup.HideCases.onPopupClick, e);
+          hide(HideCases.onPopupClick, e);
         } else {
-          hide(WUPPopup.HideCases.onOutsideClick, e);
+          hide(HideCases.onOutsideClick, e);
           wasOutsideClick = true;
           setTimeout(() => (wasOutsideClick = false), 50);
         }
@@ -191,10 +188,10 @@ export default function popupListen(
       if (!skip) {
         if (!openedEl) {
           lastActive = document.activeElement as HTMLElement;
-          show(WUPPopup.ShowCases.onClick, e);
+          show(ShowCases.onClick, e);
         } else {
           const isMeClick = openedEl === e.target || includes(e.target);
-          hide(isMeClick ? WUPPopup.HideCases.onPopupClick : WUPPopup.HideCases.onTargetClick, e);
+          hide(isMeClick ? HideCases.onPopupClick : HideCases.onTargetClick, e);
         }
       }
 
@@ -204,7 +201,7 @@ export default function popupListen(
   }
 
   // onHover
-  if (opts.showCase & WUPPopup.ShowCases.onHover) {
+  if (opts.showCase & ShowCases.onHover) {
     let timeoutId: ReturnType<typeof setTimeout> | undefined;
     const ev = (ms: number, isMouseIn: boolean, e: MouseEvent): void => {
       timeoutId && clearTimeout(timeoutId);
@@ -214,8 +211,8 @@ export default function popupListen(
           () =>
             t.isConnected && // possible when target removed via innerHTML
             (isMouseIn
-              ? show(WUPPopup.ShowCases.onHover, e) //
-              : hide(WUPPopup.HideCases.onMouseLeave, e)),
+              ? show(ShowCases.onHover, e) //
+              : hide(HideCases.onMouseLeave, e)),
           ms
         );
       else timeoutId = undefined;
@@ -226,11 +223,11 @@ export default function popupListen(
   }
 
   // onFocus
-  if (opts.showCase & WUPPopup.ShowCases.onFocus) {
+  if (opts.showCase & ShowCases.onFocus) {
     const onFocused = async (e: FocusEvent): Promise<void> => {
       if (!openedEl || debounceTimeout) {
-        preventClickAfterFocus = !!(opts.showCase & WUPPopup.ShowCases.onClick);
-        await show(WUPPopup.ShowCases.onFocus, e);
+        preventClickAfterFocus = !!(opts.showCase & ShowCases.onClick);
+        await show(ShowCases.onFocus, e);
         if (preventClickAfterFocus) {
           const rst = (): void => {
             preventClickAfterFocus = false;
@@ -249,7 +246,7 @@ export default function popupListen(
       if (openedEl) {
         const isToMe = openedEl === document.activeElement || openedEl === e.relatedTarget;
         const isToMeInside = !isToMe && includes(document.activeElement || e.relatedTarget);
-        !isToMe && !isToMeInside && (await hide(WUPPopup.HideCases.onFocusOut, e));
+        !isToMe && !isToMeInside && (await hide(HideCases.onFocusOut, e));
         /* istanbul ignore else */
         if (!openedEl) {
           openedByHover = false;
@@ -270,7 +267,7 @@ export default function popupListen(
 }
 
 popupListen.$defaults = {
-  showCase: WUPPopup.ShowCases.onClick,
+  showCase: ShowCases.onClick,
   hoverShowTimeout: 200,
   hoverHideTimeout: 500,
 };